test(aspect): Streamline target mapping integration tests
- Proper expressive names
- Less redundancy
- More coverage for using external targets
