+++ test/aspect/target_mapping/external_repo/ext_a.h
@@ -1,0 +1,5 @@
+#include "ext_b.h"
+
+int doExtA() {
+    return 42 + doExtB();
+}
