+++ test/aspect/target_mapping/external_repo/ext_b.h
@@ -1,3 +1,3 @@
-int doExt() {
+int doExtB() {
     return 42;
 }
