--- test/aspect/target_mapping/libs/c.h
@@ -1,3 +1,0 @@
-int doC() {
-    return 42;
-}
