+++ test/aspect/target_mapping/libs/mapped_dep.h
@@ -1,0 +1,6 @@
+#include "ext_a.h"
+#include "target_mapping/libs/a.h"
+
+int doMapped() {
+    return 42 + doA() + doExtA();
+}
