--- test/aspect/target_mapping/libs/with_external_dep.h
@@ -1,3 +1,0 @@
-int doWithExternalDep() {
-    return 42;
-}
