+++ test/aspect/target_mapping/mapping/BUILD
@@ -6,7 +6,10 @@ package(default_visibility = ["//target_mapping:__pkg__"])
 dwyu_make_cc_info_mapping(
     name = "map_specific_deps",
     mapping = {
-        "//target_mapping/libs:a": ["//target_mapping/libs:b"],
+        "//target_mapping/libs:mapped_dep": [
+            "//target_mapping/libs:a",
+            "@target_mapping_test_repo//:ext_a",
+        ],
     },
 )
 
@@ -14,7 +17,7 @@ dwyu_make_cc_info_mapping(
 dwyu_make_cc_info_mapping(
     name = "map_direct_deps_legacy",
     mapping = {
-        "//target_mapping/libs:a": MAP_DIRECT_DEPS,
+        "//target_mapping/libs:mapped_dep": MAP_DIRECT_DEPS,
     },
 )
 
@@ -22,34 +25,34 @@ dwyu_make_cc_info_mapping(
 dwyu_make_cc_info_mapping(
     name = "map_transitive_deps_legacy",
     mapping = {
-        "//target_mapping/libs:a": MAP_TRANSITIVE_DEPS,
+        "//target_mapping/libs:mapped_dep": MAP_TRANSITIVE_DEPS,
     },
 )
 
 dwyu_make_cc_info_mapping(
     name = "map_direct_deps_all",
     mapping = {
-        "//target_mapping/libs:a": {MAP_DIRECT_DEPS: "all"},
+        "//target_mapping/libs:mapped_dep": {MAP_DIRECT_DEPS: "all"},
     },
 )
 
 dwyu_make_cc_info_mapping(
     name = "map_transitive_deps_all",
     mapping = {
-        "//target_mapping/libs:a": {MAP_TRANSITIVE_DEPS: "all"},
+        "//target_mapping/libs:mapped_dep": {MAP_TRANSITIVE_DEPS: "all"},
     },
 )
 
 dwyu_make_cc_info_mapping(
     name = "map_direct_deps_filtered",
     mapping = {
-        "//target_mapping/libs:with_external_dep": {MAP_DIRECT_DEPS: [to_canonical("@target_mapping_test_repo//:ext")]},
+        "//target_mapping/libs:mapped_dep": {MAP_DIRECT_DEPS: [to_canonical("@target_mapping_test_repo//:ext_a")]},
     },
 )
 
 dwyu_make_cc_info_mapping(
     name = "map_transitive_deps_filtered",
     mapping = {
-        "//target_mapping/libs:with_external_dep": {MAP_TRANSITIVE_DEPS: [to_canonical("@target_mapping_test_repo//:ext")]},
+        "//target_mapping/libs:mapped_dep": {MAP_TRANSITIVE_DEPS: [to_canonical("@target_mapping_test_repo//:ext_a")]},
     },
 )
