+++ test/aspect/target_mapping/test_map_transitive_deps_all_legacy.py
@@ -7,7 +7,12 @@ from test.support.result import Result
 class TestCase(TestCaseBase):
     def execute_test_logic(self) -> Result:
         actual = self._run_dwyu(
-            target="//target_mapping:use_lib_c",
+            # Ensure using transitive and direct deps both work
+            target=[
+                "//target_mapping:use_lib_a_and_ext_a",
+                "//target_mapping:use_lib_a_and_ext_a_privately",
+                "//target_mapping:use_lib_b_and_ext_b",
+            ],
             aspect=self.choose_aspect("//target_mapping:aspect.bzl%map_transitive_deps_legacy"),
         )
