+++ test/aspect/target_mapping/test_map_transitive_deps_filtered.py
@@ -6,16 +6,16 @@ from test.support.result import Result
 
 class TestCase(TestCaseBase):
     def execute_test_logic(self) -> Result:
-        target = "//target_mapping:use_lib_c"
+        target = "//target_mapping:use_lib_b_and_ext_b"
+        # We see an error, because we mapped only one of the two used transitive dependencies
         expected = ExpectedFailure(
             ExpectedDwyuFailure(
                 target=target,
-                invalid_includes={"target_mapping/use_lib_c.cpp": ["target_mapping/libs/c.h"]},
-                unused_public_deps=["//target_mapping/libs:a"],
+                invalid_includes={"target_mapping/use_lib_b_and_ext_b.cpp": ["target_mapping/libs/b.h"]},
             )
         )
         actual = self._run_dwyu(
-            target=target, aspect=self.choose_aspect("//target_mapping:aspect.bzl%map_direct_deps_legacy")
+            target=target, aspect=self.choose_aspect("//target_mapping:aspect.bzl%map_transitive_deps_filtered")
         )
 
         return self._check_result(actual=actual, expected=expected)
