--- test/aspect/target_mapping/test_use_b_with_all_direct_deps_mapping.py
@@ -1,14 +1,0 @@
-from expected_result import ExpectedSuccess
-from test_case import TestCaseBase
-
-from test.support.result import Result
-
-
-class TestCase(TestCaseBase):
-    def execute_test_logic(self) -> Result:
-        actual = self._run_dwyu(
-            target="//target_mapping:use_lib_b",
-            aspect=self.choose_aspect("//target_mapping:aspect.bzl%map_direct_deps_all"),
-        )
-
-        return self._check_result(actual=actual, expected=ExpectedSuccess())
