--- test/aspect/target_mapping/test_use_b_with_filtered_direct_deps_mapping.py
@@ -1,21 +1,0 @@
-from expected_result import ExpectedDwyuFailure, ExpectedFailure
-from test_case import TestCaseBase
-
-from test.support.result import Result
-
-
-class TestCase(TestCaseBase):
-    def execute_test_logic(self) -> Result:
-        target = "//target_mapping:use_b_via_with_external_dep"
-        expected = ExpectedFailure(
-            ExpectedDwyuFailure(
-                target=target,
-                invalid_includes={"target_mapping/use_b_via_with_external_dep.cpp": ["target_mapping/libs/b.h"]},
-                unused_public_deps=["//target_mapping/libs:with_external_dep"],
-            )
-        )
-        actual = self._run_dwyu(
-            target=target, aspect=self.choose_aspect("//target_mapping:aspect.bzl%map_direct_deps_filtered")
-        )
-
-        return self._check_result(actual=actual, expected=expected)
