--- test/aspect/target_mapping/use_c_via_with_external_dep.cpp
@@ -1,5 +1,0 @@
-#include "target_mapping/libs/c.h"
-
-int main() {
-    return doC();
-}
