--- test/aspect/target_mapping/use_ext_via_with_external_dep.cpp
@@ -1,5 +1,0 @@
-#include "ext.h"
-
-int main() {
-    return doExt();
-}
