+++ test/aspect/target_mapping/use_lib_a_and_ext_a.cpp
@@ -1,0 +1,6 @@
+#include "ext_a.h"
+#include "target_mapping/libs/a.h"
+
+int main() {
+    return doA() + doExtA();
+}
