--- test/aspect/target_mapping/use_lib_b.cpp
@@ -1,5 +1,0 @@
-#include "target_mapping/libs/b.h"
-
-int main() {
-    return doB();
-}
