+++ test/aspect/target_mapping/use_lib_b_and_ext_b.cpp
@@ -1,5 +1,6 @@
+#include "ext_b.h"
 #include "target_mapping/libs/b.h"
 
 int main() {
-    return doB();
+    return doB() + doExtB();
 }
