--- test/aspect/target_mapping/use_lib_c.cpp
@@ -1,5 +1,0 @@
-#include "target_mapping/libs/c.h"
-
-int main() {
-    return doC();
-}
